kyc_backend/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
Pillow
pypdfium2
rapidocr-onnxruntime
rapidfuzz
python-multipart
numpy
